+++ src/reflect/scala/reflect/io/FileOperationException.scala
@@ -7,7 +7,7 @@
 \*                                                                      */
 
 
-package scala.tools.nsc
+package scala.reflect
 package io
-
+/** ''Note:  This library is considered experimental and should not be used unless you know what you are doing.'' */
 case class FileOperationException(msg: String) extends RuntimeException(msg)
