+++ src/reflect/scala/reflect/io/NoAbstractFile.scala
@@ -3,7 +3,7 @@
  * @author  Paul Phillips
  */
 
-package scala.tools.nsc
+package scala.reflect
 package io
 
 import java.io.InputStream
@@ -11,6 +11,8 @@ import java.io.{ File => JFile }
 
 /** A distinguished object so you can avoid both null
  *  and Option.
+ *  
+ *  ''Note:  This library is considered experimental and should not be used unless you know what you are doing.''
  */
 object NoAbstractFile extends AbstractFile {
   def absolute: AbstractFile = this
